Move test_utils around so others can depend on it
