+++ src/rust/e2e-tests/Cargo.toml
@@ -6,7 +6,7 @@ edition = "2021"
 [features]
 integration_tests = []
 
-[dev-dependencies]
+[dependencies]
 async-trait = "0.1"
 bytes = "1.1"
 clap = { version = "3.0", default_features = false, features = ["std", "env"] }
