+++ src/rust/e2e-tests/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod test_fixtures;
+pub mod test_utils;
